+++ BuldingMapper/FormResults/FormResult.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BuildingMapper.FormResults
+{
+    public abstract class FormResult
+    {
+        public DialogResult DialogResult { get; set; }
+    }
+}
