+++ BuldingMapper/FormResults/SaveWarningFormResult.cs
@@ -4,10 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BuildingMapper
+namespace BuildingMapper.FormResults
 {
-    public abstract class FormResult
+    internal class SaveWarningFormResult
     {
-        public DialogResult DialogResult { get; set; }
     }
 }
