+++ BuldingMapper/RoomEditorForm.cs
@@ -7,6 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using BuildingMapper.FormResults;
 
 namespace BuildingMapper
 {
